LC-unique-binary-search-trees [Runtime: N/A; Memory: N/A]

+++ math/unique-binary-search-trees.cpp
@@ -4,9 +4,9 @@ public:
         if(n==1){
             return 1;
         }
-        int num = 1;
-        int d1 = 1;
-        int d2 = 1;
+        long long num = 1;
+        long long d1 = 1;
+        long long d2 = 1;
 
         for(int i = 1; i < n+1; i++){
             d1 = d1 * i;
